% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management.R
\name{add_user}
\alias{add_user}
\title{Add User}
\usage{
add_user(domain, group, email)
}
\arguments{
\item{domain}{Required. A GSuite domain: 'somedomain.org'}

\item{group}{Required. The name of a Google Group. A listing of groups in your domain can be obtained by running list_groups(domain = 'mydomain.org')}

\item{email}{Required. The Gmail Account of the user that you would like to add to the group.}
}
\value{
JSON object detailing successful user addition attributes.
}
\description{
Add a user to a Google Group
}
