% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_group.R
\name{create_group}
\alias{create_group}
\title{Create Group}
\usage{
create_group(email, name = NULL, description = NULL)
}
\arguments{
\item{email}{Required. Email address to assign to your group. 'test@gsuitedomain.org'}

\item{name}{Optional. Provide a descriptive name for your group.}

\item{description}{Optional. Provide a description for your group.}
}
\value{
JSON object with information about the newly created group.
}
\description{
Create a group in the a GSuite domain.
}
