% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_management.R
\name{delete_group}
\alias{delete_group}
\title{Delete Group}
\usage{
delete_group(domain, group)
}
\arguments{
\item{domain}{A GSuite domain: 'somedomain.org'}

\item{group}{The name of a Google Group. A listing of groups in your domain can be obtained by running list_groups(domain = 'mydomain.org')}
}
\value{
If successful, an empty JSON responose will be received.
}
\description{
Delete a group in a GSuite domain.
}
