% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_access_token.R
\name{get_access_token}
\alias{get_access_token}
\title{Get Access Token}
\usage{
get_access_token(cached_credentials = "~/.adminSDK_cache.rds")
}
\arguments{
\item{cached_credentials}{Location to read from/store credentials.}
}
\value{
Access token used to perform AdminSDK tasks
}
\description{
Get Access Token
}
