% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_id.R
\name{get_group_id}
\alias{get_group_id}
\title{Get Group ID}
\usage{
get_group_id(domain, group)
}
\arguments{
\item{domain}{A GSuite domain: 'somedomain.org'}

\item{group}{The name of a Google Group. A listing of groups in your domain can be obtained by running list_groups(domain = 'mydomain.org')}
}
\value{
Google Group ID
}
\description{
Retrieve the Group ID associated with a Group Name
}
\keyword{internal}
