% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_groups.R
\name{list_groups}
\alias{list_groups}
\title{List Groups}
\usage{
list_groups(domain)
}
\arguments{
\item{domain}{A Google domain}
}
\value{

}
\description{
Retrieve a listing of all groups in the learnclinicalda
}
