% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_user_info.R
\name{list_user_info}
\alias{list_user_info}
\title{List User Info}
\usage{
list_user_info(domain, group, memberKey)
}
\arguments{
\item{domain}{A GSuite domain: 'somedomain.org'}

\item{group}{The name of a Google Group. A listing of groups in your domain can be obtained by running list_groups(domain = 'mydomain.org')}

\item{memberKey}{The user's unique ID}
}
\value{
JSON object with directory information pertaining to the user's unique ID.
}
\description{
Retrieve a user specific directory info, such as primary email address.
}
