% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management.R
\name{remove_user}
\alias{remove_user}
\title{Remove User}
\usage{
remove_user(domain, group, email)
}
\arguments{
\item{domain}{A GSuite domain: 'somedomain.org'}

\item{group}{The name of a Google Group. A listing of groups in your domain can be obtained by running list_groups(domain = 'mydomain.org')}

\item{email}{The Gmail Account of the user that you would like to add to the group.}
}
\value{

}
\description{
Remove a user from a Google Group
}
