% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_groups.R
\name{list_groups}
\alias{list_groups}
\title{List Groups}
\usage{
list_groups(domain)
}
\arguments{
\item{domain}{Required. A GSuite domain: 'somedomain.org'}
}
\value{
JSON object with information about all groups in the specified GSuite domain.
}
\description{
Retrieve a listing of all groups in the specified GSuite domain.
}
